Add javadoc and license

+++ src/main/java/VRDownloader.java
@@ -3,6 +3,9 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.net.Socket;
 
+/**
+ * For VRPlayer to download manifest and video segments in a separate thread.
+ */
 public class VRDownloader {
     private String host;
     private int port;
@@ -16,7 +19,13 @@ public class VRDownloader {
 
     /**
      * Setup for downloading a video segment.
-     * @param port
+     *
+     * @param host     to the VRServer.
+     * @param port     to the VRServer.
+     * @param path     for the manifest file.
+     * @param filename of the video segment.
+     * @param snb      the segment number of the video segment.
+     * @param segLen   the file size of the video segment.
      */
     public VRDownloader(String host, int port, String path, String filename, int snb, int segLen) {
         // init instance variables
@@ -32,9 +41,10 @@ public class VRDownloader {
 
     /**
      * Setup for downloading manifest file.
-     * @param host
-     * @param port
-     * @param path
+     *
+     * @param host to the VRServer.
+     * @param port to the VRServer.
+     * @param path for the manifest file.
      */
     public VRDownloader(String host, int port, String path) {
         // init instance variables
@@ -63,8 +73,9 @@ public class VRDownloader {
 
     /**
      * Save manifest from VRServer.
-     * @param clientSock
-     * @throws IOException
+     *
+     * @param clientSock the socket for video downloader to save manifest file.
+     * @throws IOException when dataInputStream fails to read or fileOutputStream fails to write.
      */
     private void saveManifest(Socket clientSock) throws IOException {
         DataInputStream dis = new DataInputStream(clientSock.getInputStream());
@@ -74,7 +85,7 @@ public class VRDownloader {
         int read = 0;
         int totalRead = 0;
         int remaining = 1402;   // TODO how to know the manifest file size when we do not have manifest?
-        while((read = dis.read(buffer, 0, Math.min(buffer.length, remaining))) > 0) {
+        while ((read = dis.read(buffer, 0, Math.min(buffer.length, remaining))) > 0) {
             totalRead += read;
             remaining -= read;
             fos.write(buffer, 0, read);
@@ -86,8 +97,9 @@ public class VRDownloader {
 
     /**
      * Save video segment from VRServer.
+     *
      * @param clientSock
-     * @throws IOException
+     * @throws IOException when dataInputStream fails to read or fileOutputStream fails to write.
      */
     private void saveVideoSegment(Socket clientSock, String path, String filename, int snb, int segLen) throws IOException {
         DataInputStream dis = new DataInputStream(clientSock.getInputStream());
@@ -95,10 +107,10 @@ public class VRDownloader {
         byte[] buffer = new byte[4096];
         int read = 0;
         int totalRead = 0;
-        int remaining = (int) segLen;
+        int remaining = segLen;
 
-        System.out.println("Saving " + Utilities.getSegmentName("",filename, snb));
-        while((read = dis.read(buffer, 0, Math.min(buffer.length, remaining))) > 0) {
+        System.out.println("Saving " + Utilities.getSegmentName("", filename, snb));
+        while ((read = dis.read(buffer, 0, Math.min(buffer.length, remaining))) > 0) {
             totalRead += read;
             remaining -= read;
             fos.write(buffer, 0, read);
